+++ tests/testthat/test-with_deduped.R
@@ -2,7 +2,10 @@ test_that("with_deduped(f(x)) runs only on deduplicated values", {
   get_ncalls <- function() as.integer(Sys.getenv("ncalls"))
   increment_ncalls <- function() Sys.setenv(ncalls = get_ncalls() + 1L)
 
-  f <- function(ii) for (i in ii) increment_ncalls()
+  f <- function(ii) {
+    for (i in ii) increment_ncalls()
+    ii
+  }
 
   withr::with_envvar(
     c(ncalls = 0),
@@ -23,12 +26,26 @@ test_that("with_deduped(f(x)) runs only on deduplicated values", {
   )
 })
 
-test_that("with_deduped(f(x, y)) is the same as deduped(f)(x, y)", {
+test_that("with_deduped(f(x, y)) is the same as deduped(f)(x, y) (int)", {
   x <- c(1, 3, 1, 2, 2, 1)
   increment_by_n <- \(x, n) x + n
   inc <- 3
   expect_equal(
-    increment_by_n(x, inc) |> with_deduped(),
-    deduped(increment_by_n)(x, inc)
+    deduped(increment_by_n)(x, inc),
+    expect_no_warning(increment_by_n(x, inc) |> with_deduped())
+  )
+
+  # Named arguments
+  expect_equal(
+    deduped(increment_by_n)(x, n=inc),
+    expect_no_warning(increment_by_n(x, n=inc) |> with_deduped())
+  )
+})
+
+test_that("with_deduped(f(x, y)) is the same as deduped(f)(x, y) (char)", {
+  x <- LETTERS[c(1, 3, 1, 2, 2, 1)]
+  expect_equal(
+    deduped(paste0)(x, "X"),
+    expect_no_warning(paste0(x, "X") |> with_deduped())
   )
 })